chore(sdk): dedicated CI workflow
