+++ sdk/packages/react/test/assets/SolverNetInbox.json
