+++ sdk/packages/react/test/assets/SolverNetMiddleman.json
